Move function into class
